readmitiq/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
redis==5.1.1
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==24.4.0

# ML Core
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.2
lightgbm==4.5.0
optuna==3.6.1
imbalanced-learn==0.12.4

# Explainability
shap==0.46.0

# Deep Learning (optional — comment out if CUDA unavailable)
# torch==2.4.1
# torchvision==0.19.1

# Utilities
python-multipart==0.0.12
httpx==0.27.2
aiofiles==24.1.0
